Introduced InvocationMonitor
I pulled out this logic from the InvocationRegistry. The InvocationRegistry is now a dumb container
for invocations. And the monitoring (checking for timeouts, member leaving the cluster) and responding
to it is a responsibility of the InvocationMonitor.

So it is a seperation of concerns.

One of the reasons, apart from having cleaner code, to have this seperation, is that in 3.7 I want to
experiment with different types of invocation-storage. Currently we use a CHM for it but it generates
quite a lot of waste.