Fixed Operation.toString
The Operation has a nice toString method which is quite informative, unfortunately it was
overridden on quite a few placed by one that doesn't include all this information.

This pr fixes that by replacing all Operation.toString methods by a better one that
can be glued into the one of the  operation

Also at important operations where the name of the operation is mentioned, the new
toString(StringBuffer) has been added so that we now get to see which data-structures
these operations are working on.